% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population for every UN country.}
\format{
a data frame with 235 observations
\describe{
  \item{CountryName}{Name of country.}
  \item{1950-2020}{Population number for each year.}
}
}
\source{
\url{http://creativecommons.org/licenses/by/3.0/igo/}
}
\usage{
WorldPopulation
}
\description{
World Population for every UN country.
}
\keyword{datasets}
